mergeConstraintExtractino
